backend/requirements.txt
python-gitlab
flask